added different icons for different device types
- moved converters into different files
- removed console output
- added static converter to App.xaml
